move verified org loader tests
