+++ api-js/src/loaders/verified-organizations/__tests__/load-verified-org-conn-by-domain-id.test.js
@@ -5,17 +5,17 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { cleanseInput } = require('../validators')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { cleanseInput } = require('../../../validators')
 const {
-  verifiedOrgLoaderConnections,
+  verifiedOrgLoaderConnectionsByDomainId,
   verifiedOrgLoaderByKey,
-} = require('../loaders')
+} = require('../..')
 
 describe('given the load organizations connection function', () => {
-  let query, drop, truncate, migrate, collections, org, orgTwo, i18n
+  let query, drop, truncate, migrate, collections, org, orgTwo, domain, i18n
 
   let consoleOutput = []
   const mockedError = (output) => consoleOutput.push(output)
@@ -105,6 +105,18 @@ describe('given the load organizations connection function', () => {
         },
       },
     })
+    domain = await collections.domains.save({
+      domain: 'test.domain.gc.ca',
+      slug: 'test-domain-gc-ca',
+    })
+    await collections.claims.save({
+      _from: org._id,
+      _to: domain._id,
+    })
+    await collections.claims.save({
+      _from: orgTwo._id,
+      _to: domain._id,
+    })
     consoleOutput = []
   })
 
@@ -127,7 +139,7 @@ describe('given the load organizations connection function', () => {
     describe('given a successful load', () => {
       describe('using no cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -138,6 +150,7 @@ describe('given the load organizations connection function', () => {
             first: 5,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -188,7 +201,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using after cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -206,6 +219,7 @@ describe('given the load organizations connection function', () => {
             after: toGlobalId('verifiedOrganizations', expectedOrgs[0].id),
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -241,7 +255,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using before cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -259,6 +273,7 @@ describe('given the load organizations connection function', () => {
             before: toGlobalId('verifiedOrganizations', expectedOrgs[1].id),
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -294,7 +309,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using first limit', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -311,6 +326,7 @@ describe('given the load organizations connection function', () => {
             first: 1,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -346,7 +362,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using last limit', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -363,6 +379,7 @@ describe('given the load organizations connection function', () => {
             last: 1,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -398,8 +415,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('no organizations are found', () => {
         it('returns empty structure', async () => {
-          await truncate()
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -432,7 +448,7 @@ describe('given the load organizations connection function', () => {
     describe('given an unsuccessful load', () => {
       describe('limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -441,7 +457,7 @@ describe('given the load organizations connection function', () => {
 
           try {
             const connectionArgs = {}
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(
               new Error(
@@ -451,13 +467,13 @@ describe('given the load organizations connection function', () => {
           }
 
           expect(consoleOutput).toEqual([
-            'User did not have either `first` or `last` arguments set for: verifiedOrgLoaderConnections.',
+            'User did not have either `first` or `last` arguments set for: verifiedOrgLoaderConnectionsByDomainId.',
           ])
         })
       })
       describe('user has first and last arguments set at the same time', () => {
         it('returns an error message', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -469,7 +485,7 @@ describe('given the load organizations connection function', () => {
               first: 1,
               last: 1,
             }
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(
               new Error(
@@ -479,14 +495,14 @@ describe('given the load organizations connection function', () => {
           }
 
           expect(consoleOutput).toEqual([
-            'User attempted to have `first` and `last` arguments set for: verifiedOrgLoaderConnections.',
+            'User attempted to have `first` and `last` arguments set for: verifiedOrgLoaderConnectionsByDomainId.',
           ])
         })
       })
       describe('limits are set below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'en',
               cleanseInput,
@@ -498,6 +514,7 @@ describe('given the load organizations connection function', () => {
                 first: -1,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -509,13 +526,13 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `first` set below zero for: verifiedOrgLoaderConnections.',
+              'User attempted to have `first` set below zero for: verifiedOrgLoaderConnectionsByDomainId.',
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'en',
               cleanseInput,
@@ -527,6 +544,7 @@ describe('given the load organizations connection function', () => {
                 last: -1,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -538,7 +556,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `last` set below zero for: verifiedOrgLoaderConnections.',
+              'User attempted to have `last` set below zero for: verifiedOrgLoaderConnectionsByDomainId.',
             ])
           })
         })
@@ -546,7 +564,7 @@ describe('given the load organizations connection function', () => {
       describe('limits are set above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'en',
               cleanseInput,
@@ -558,6 +576,7 @@ describe('given the load organizations connection function', () => {
                 first: 101,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -569,13 +588,13 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `first` to 101 for: verifiedOrgLoaderConnections.',
+              'User attempted to have `first` to 101 for: verifiedOrgLoaderConnectionsByDomainId.',
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'en',
               cleanseInput,
@@ -587,6 +606,7 @@ describe('given the load organizations connection function', () => {
                 last: 101,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -598,7 +618,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `last` to 101 for: verifiedOrgLoaderConnections.',
+              'User attempted to have `last` to 101 for: verifiedOrgLoaderConnectionsByDomainId.',
             ])
           })
         })
@@ -609,7 +629,7 @@ describe('given the load organizations connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = verifiedOrgLoaderConnections(
+              const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
                 query,
                 'en',
                 cleanseInput,
@@ -632,7 +652,7 @@ describe('given the load organizations connection function', () => {
                 )
               }
               expect(consoleOutput).toEqual([
-                `User attempted to have \`first\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnections.`,
+                `User attempted to have \`first\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -642,7 +662,7 @@ describe('given the load organizations connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = verifiedOrgLoaderConnections(
+              const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
                 query,
                 'en',
                 cleanseInput,
@@ -665,7 +685,7 @@ describe('given the load organizations connection function', () => {
                 )
               }
               expect(consoleOutput).toEqual([
-                `User attempted to have \`last\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnections.`,
+                `User attempted to have \`last\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -679,7 +699,7 @@ describe('given the load organizations connection function', () => {
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
 
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'en',
             cleanseInput,
@@ -690,7 +710,7 @@ describe('given the load organizations connection function', () => {
             const connectionArgs = {
               first: 5,
             }
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(
               new Error(
@@ -700,7 +720,7 @@ describe('given the load organizations connection function', () => {
           }
 
           expect(consoleOutput).toEqual([
-            `Database error occurred while user was trying to gather orgs in verifiedOrgLoaderConnections, error: Error: Database error occurred.`,
+            `Database error occurred while user was trying to gather orgs in verifiedOrgLoaderConnectionsByDomainId, error: Error: Database error occurred.`,
           ])
         })
       })
@@ -714,7 +734,7 @@ describe('given the load organizations connection function', () => {
             }
             const query = jest.fn().mockReturnValueOnce(cursor)
 
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'en',
               cleanseInput,
@@ -726,6 +746,7 @@ describe('given the load organizations connection function', () => {
                 first: 5,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -737,7 +758,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              `Cursor error occurred while user was trying to gather orgs in verifiedOrgLoaderConnections, error: Error: Cursor error occurred.`,
+              `Cursor error occurred while user was trying to gather orgs in verifiedOrgLoaderConnectionsByDomainId, error: Error: Cursor error occurred.`,
             ])
           })
         })
@@ -759,7 +780,7 @@ describe('given the load organizations connection function', () => {
     describe('given a successful load', () => {
       describe('using no cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -770,6 +791,7 @@ describe('given the load organizations connection function', () => {
             first: 5,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -820,7 +842,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using after cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -838,6 +860,7 @@ describe('given the load organizations connection function', () => {
             after: toGlobalId('verifiedOrganizations', expectedOrgs[0].id),
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -873,7 +896,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using before cursor', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -891,6 +914,7 @@ describe('given the load organizations connection function', () => {
             before: toGlobalId('verifiedOrganizations', expectedOrgs[1].id),
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -926,7 +950,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using first limit', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -943,6 +967,7 @@ describe('given the load organizations connection function', () => {
             first: 1,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -978,7 +1003,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('using last limit', () => {
         it('returns an organization', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -995,6 +1020,7 @@ describe('given the load organizations connection function', () => {
             last: 1,
           }
           const orgs = await connectionLoader({
+            domainId: domain._id,
             ...connectionArgs,
           })
 
@@ -1030,8 +1056,7 @@ describe('given the load organizations connection function', () => {
       })
       describe('no organizations are found', () => {
         it('returns empty structure', async () => {
-          await truncate()
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -1042,6 +1067,7 @@ describe('given the load organizations connection function', () => {
             last: 1,
           }
           const orgs = await connectionLoader({
+            domainId: 'domains/1',
             ...connectionArgs,
           })
 
@@ -1063,7 +1089,7 @@ describe('given the load organizations connection function', () => {
     describe('given an unsuccessful load', () => {
       describe('limits are not set', () => {
         it('returns an error message', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -1072,19 +1098,19 @@ describe('given the load organizations connection function', () => {
 
           try {
             const connectionArgs = {}
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
 
           expect(consoleOutput).toEqual([
-            'User did not have either `first` or `last` arguments set for: verifiedOrgLoaderConnections.',
+            `User did not have either \`first\` or \`last\` arguments set for: verifiedOrgLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('user has first and last arguments set at the same time', () => {
         it('returns an error message', async () => {
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -1096,20 +1122,20 @@ describe('given the load organizations connection function', () => {
               first: 1,
               last: 1,
             }
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
 
           expect(consoleOutput).toEqual([
-            'User attempted to have `first` and `last` arguments set for: verifiedOrgLoaderConnections.',
+            `User attempted to have \`first\` and \`last\` arguments set for: verifiedOrgLoaderConnectionsByDomainId.`,
           ])
         })
       })
       describe('limits are set below minimum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'fr',
               cleanseInput,
@@ -1121,6 +1147,7 @@ describe('given the load organizations connection function', () => {
                 first: -1,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -1128,13 +1155,13 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `first` set below zero for: verifiedOrgLoaderConnections.',
+              `User attempted to have \`first\` set below zero for: verifiedOrgLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'fr',
               cleanseInput,
@@ -1146,6 +1173,7 @@ describe('given the load organizations connection function', () => {
                 last: -1,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -1153,7 +1181,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `last` set below zero for: verifiedOrgLoaderConnections.',
+              `User attempted to have \`last\` set below zero for: verifiedOrgLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -1161,7 +1189,7 @@ describe('given the load organizations connection function', () => {
       describe('limits are set above maximum', () => {
         describe('first is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'fr',
               cleanseInput,
@@ -1173,6 +1201,7 @@ describe('given the load organizations connection function', () => {
                 first: 101,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -1180,13 +1209,13 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `first` to 101 for: verifiedOrgLoaderConnections.',
+              `User attempted to have \`first\` to 101 for: verifiedOrgLoaderConnectionsByDomainId.`,
             ])
           })
         })
         describe('last is set', () => {
           it('returns an error message', async () => {
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'fr',
               cleanseInput,
@@ -1198,6 +1227,7 @@ describe('given the load organizations connection function', () => {
                 last: 101,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -1205,7 +1235,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              'User attempted to have `last` to 101 for: verifiedOrgLoaderConnections.',
+              `User attempted to have \`last\` to 101 for: verifiedOrgLoaderConnectionsByDomainId.`,
             ])
           })
         })
@@ -1216,7 +1246,7 @@ describe('given the load organizations connection function', () => {
             it(`returns an error when first set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = verifiedOrgLoaderConnections(
+              const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
                 query,
                 'fr',
                 cleanseInput,
@@ -1235,7 +1265,7 @@ describe('given the load organizations connection function', () => {
                 expect(err).toEqual(new Error(`todo`))
               }
               expect(consoleOutput).toEqual([
-                `User attempted to have \`first\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnections.`,
+                `User attempted to have \`first\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1245,7 +1275,7 @@ describe('given the load organizations connection function', () => {
             it(`returns an error when last set to ${stringify(
               invalidInput,
             )}`, async () => {
-              const connectionLoader = verifiedOrgLoaderConnections(
+              const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
                 query,
                 'fr',
                 cleanseInput,
@@ -1264,7 +1294,7 @@ describe('given the load organizations connection function', () => {
                 expect(err).toEqual(new Error(`todo`))
               }
               expect(consoleOutput).toEqual([
-                `User attempted to have \`last\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnections.`,
+                `User attempted to have \`last\` set as a ${typeof invalidInput} for: verifiedOrgLoaderConnectionsByDomainId.`,
               ])
             })
           })
@@ -1278,7 +1308,7 @@ describe('given the load organizations connection function', () => {
             .fn()
             .mockRejectedValue(new Error('Database error occurred.'))
 
-          const connectionLoader = verifiedOrgLoaderConnections(
+          const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
             query,
             'fr',
             cleanseInput,
@@ -1289,13 +1319,13 @@ describe('given the load organizations connection function', () => {
             const connectionArgs = {
               first: 5,
             }
-            await connectionLoader({ ...connectionArgs })
+            await connectionLoader({ domainId: domain._id, ...connectionArgs })
           } catch (err) {
             expect(err).toEqual(new Error('todo'))
           }
 
           expect(consoleOutput).toEqual([
-            `Database error occurred while user was trying to gather orgs in verifiedOrgLoaderConnections, error: Error: Database error occurred.`,
+            `Database error occurred while user was trying to gather orgs in verifiedOrgLoaderConnectionsByDomainId, error: Error: Database error occurred.`,
           ])
         })
       })
@@ -1309,7 +1339,7 @@ describe('given the load organizations connection function', () => {
             }
             const query = jest.fn().mockReturnValueOnce(cursor)
 
-            const connectionLoader = verifiedOrgLoaderConnections(
+            const connectionLoader = verifiedOrgLoaderConnectionsByDomainId(
               query,
               'fr',
               cleanseInput,
@@ -1321,6 +1351,7 @@ describe('given the load organizations connection function', () => {
                 first: 5,
               }
               await connectionLoader({
+                domainId: domain._id,
                 ...connectionArgs,
               })
             } catch (err) {
@@ -1328,7 +1359,7 @@ describe('given the load organizations connection function', () => {
             }
 
             expect(consoleOutput).toEqual([
-              `Cursor error occurred while user was trying to gather orgs in verifiedOrgLoaderConnections, error: Error: Cursor error occurred.`,
+              `Cursor error occurred while user was trying to gather orgs in verifiedOrgLoaderConnectionsByDomainId, error: Error: Cursor error occurred.`,
             ])
           })
         })
