+++ api-js/src/loaders/verified-organizations/__tests__/load-verified-organization-by-slug.test.js
@@ -3,10 +3,10 @@ const { DB_PASS: rootPass, DB_URL: url } = process.env
 const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { setupI18n } = require('@lingui/core')
 
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
-const { makeMigrations } = require('../../migrations')
-const { verifiedOrgLoaderBySlug } = require('../loaders')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { verifiedOrgLoaderBySlug } = require('../..')
 
 describe('given a verifiedOrgLoaderBySlug dataloader', () => {
   let query, drop, truncate, migrate, collections, i18n
